+++ src/code/DCUIBubble.h
@@ -96,16 +96,6 @@
 /// ---------------------------------------
 
 /**
- Called by the DCUIRating control to move this control to it's new horizontal position. 
- 
- Or if you are using a
- bubble over another control, ensure that the touch events of that control call this.
- 
- @param aTouch the touch event that triggered the move.
-- (void) alignWithTough:(UITouch *)aTouch;
- */
-
-/**
  * Called by the parent control to position the bubble above it and to the right by a specified offset.
  * @param view the parent view.
  * @param offset how far along the control to display the bubble in pixels.